Move getItem etc. to Abstract
We take the hint from the naming PyObject_* in CPython that these
methods are not sequence API. In Jython we do not need
PySequence.getItem etc. distinct from PyObject equivalents.
