+++ rt4core/src/main/java/uk/co/farowl/vsj4/core/PySequence.java
@@ -15,7 +15,6 @@ import java.util.stream.StreamSupport;
 
 import uk.co.farowl.vsj4.core.PySlice.Indices;
 import uk.co.farowl.vsj4.core.PyUtil.NoConversion;
-import uk.co.farowl.vsj4.internal.EmptyException;
 import uk.co.farowl.vsj4.kernel.BaseType;
 import uk.co.farowl.vsj4.kernel.KernelTypeFlag;
 import uk.co.farowl.vsj4.kernel.Representation;
@@ -28,7 +27,7 @@ import uk.co.farowl.vsj4.types.TypeFlag;
  */
 public class PySequence extends Abstract {
 
-    PySequence() {}   // only static methods here
+    private PySequence() {}   // only static methods here
 
     /**
      * Return whether the object (its type, rather) provides the
@@ -74,141 +73,6 @@ public class PySequence extends Abstract {
     public static Object concat(Object v, Object w) throws Throwable {
         // There is no equivalent slot to sq_concat
         return PyNumber.add(v, w);
-    }
-
-    /**
-     * {@code o[key]} with Python semantics, where {@code o} may be a
-     * mapping or a sequence.
-     *
-     * @param o object to operate on
-     * @param key index
-     * @return {@code o[key]}
-     * @throws PyBaseException ({@link PyExc#TypeError TypeError}) when
-     *     {@code o} does not allow subscripting
-     * @throws Throwable from invoked method implementations
-     */
-    // Compare CPython PyObject_GetItem in abstract.c
-    public static Object getItem(Object o, Object key)
-            throws Throwable {
-        // Decisions are based on types of o and key
-        try {
-            return representation(o).op_getitem().invokeExact(o, key);
-        } catch (EmptyException e) {
-            throw typeError(NOT_SUBSCRIPTABLE, o);
-        }
-    }
-
-    /**
-     * {@code o[i1:12]} with Python semantics, where {@code o} must be a
-     * sequence. Receiving objects will normally interpret indices as
-     * end-relative, and bounded to the sequence length.
-     *
-     * @param o sequence to operate on
-     * @param i1 index of first item in slice
-     * @param i2 index of first item not in slice
-     * @return {@code o[i1:i2]}
-     * @throws PyBaseException ({@link PyExc#TypeError TypeError}) when
-     *     {@code o} does not allow subscripting
-     * @throws Throwable from invoked method implementations
-     */
-    // Compare CPython PyObject_GetItem in abstract.c
-    public static Object getSlice(Object o, int i1, int i2)
-            throws Throwable {
-        Object key = new PySlice(i1, i2);
-        // Decisions are based on type of o and known type of key
-        try {
-            return representation(o).op_getitem().invokeExact(o, key);
-        } catch (EmptyException e) {
-            throw typeError(NOT_SLICEABLE, o);
-        }
-    }
-
-    /**
-     * {@code o[key] = value} with Python semantics, where {@code o} may
-     * be a mapping or a sequence.
-     *
-     * @param o object to operate on
-     * @param key index
-     * @param value to put at index
-     * @throws PyBaseException ({@link PyExc#TypeError TypeError}) when
-     *     {@code o} does not allow subscripting
-     * @throws Throwable from invoked method implementations
-     */
-    // Compare CPython PyObject_SetItem in abstract.c
-    public static void setItem(Object o, Object key, Object value)
-            throws Throwable {
-        // Decisions are based on types of o and key
-        try {
-            representation(o).op_setitem().invokeExact(o, key, value);
-        } catch (EmptyException e) {
-            throw typeError(DOES_NOT_SUPPORT_ITEM, o, "assignment");
-        }
-    }
-
-    /**
-     * {@code o[i1:12] = value} with Python semantics, where {@code o} may
-     * be a mapping or a sequence.
-     *
-     * @param o object to operate on
-     * @param i1 index of first item in slice
-     * @param i2 index of first item not in slice
-     * @param value to replace the slice
-     * @throws PyBaseException ({@link PyExc#TypeError TypeError}) when
-     *     {@code o} does not allow subscripting
-     * @throws Throwable from invoked method implementations
-     */
-    // Compare CPython PyObject_SetSlice in abstract.c
-    public static void setSlice(Object o, int i1, int i2, Object value)
-            throws Throwable {
-        Object key = new PySlice(i1, i2);
-        // Decisions are based on types of o and key
-        try {
-            representation(o).op_setitem().invokeExact(o, key, value);
-        } catch (EmptyException e) {
-            throw typeError(DOES_NOT_SUPPORT_ITEM, o, "assignment");
-        }
-    }
-
-    /**
-     * {@code del o[key]} with Python semantics, where {@code o} may be
-     * a mapping or a sequence.
-     *
-     * @param o object to operate on
-     * @param key index at which to delete element
-     * @throws PyBaseException ({@link PyExc#TypeError TypeError}) when
-     *     {@code o} does not allow subscripting
-     * @throws Throwable from invoked method implementations
-     */
-    // Compare CPython PyObject_DelItem in abstract.c
-    public static void delItem(Object o, Object key) throws Throwable {
-        // Decisions are based on types of o and key
-        try {
-            representation(o).op_delitem().invokeExact(o, key);
-        } catch (EmptyException e) {
-            throw typeError(DOES_NOT_SUPPORT_ITEM, o, "deletion");
-        }
-    }
-
-    /**
-     * {@code del o[i1:12]} with Python semantics, where {@code o} may be
-     * a mapping or a sequence.
-     *
-     * @param o object to operate on
-     * @param i1 index of first item in slice
-     * @param i2 index of first item not in slice
-     * @throws PyBaseException ({@link PyExc#TypeError TypeError}) when
-     *     {@code o} does not allow subscripting
-     * @throws Throwable from invoked method implementations
-     */
-    // Compare CPython PyObject_DelSlice in abstract.c
-    public static void delSlice(Object o, int i1, int i2) throws Throwable {
-        Object key = new PySlice(i1, i2);
-        // Decisions are based on types of o and key
-        try {
-            representation(o).op_delitem().invokeExact(o, key);
-        } catch (EmptyException e) {
-            throw typeError(DOES_NOT_SUPPORT_ITEM, o, "deletion");
-        }
     }
 
     /**
@@ -432,15 +296,6 @@ public class PySequence extends Abstract {
         }
         return false;
     }
-
-    // Strings for constructing error messages ------------------------
-
-    private static final String NOT_SUBSCRIPTABLE =
-            "'%.200s' object is not subscriptable";
-    private static final String NOT_SLICEABLE =
-            "'%.200s' object is unsliceable";
-    static final String DOES_NOT_SUPPORT_ITEM =
-            "'%.200s' object does not support item %s";
 
     // Classes supporting implementations of sequence types -----------
